Exiting operator honoring deferred calls
The `main` function does not allow calling the deferred function in case of an error. For example, I would like to add something like this to the main function:

``go
defer mymetrics.Unregister()
```
The preceding line will not be called if any error occurs, because all errors are handled this way:

```go
if err != nil {
	setupLog.Error(err, "some message")
	os.Exit(1)
}
```

`os.Exit()` will exit the program immediately and deferred functions will be ignored.

I've prepared a PR that would terminate the program perhaps a bit more elegantly.

Simplified, it would be to implement something like this:

```go
func main() {
    var exitCode = 1
    defer func() {
        os.Exit(exitCode)
    }()

    defer f2()
    defer f1()

    err := .....
    if err != nil {
       // executes all previous deferreds in order f1(), f2(), os.Exit(1)
       return
    }

    err = ....
    if err != nil {
       // executes all previous deferreds in order f1(), f2(), os.Exit(1)
       return
    }

    exitCode = 0
    // All went well,
    // executes all previous deferreds in order f1(), f2(), os.Exit(0)
}
```

Signed-off-by: kuritka <kuritka@gmail.com>